fix(Country): cambio del tipo de area en models

+++ cr-pi-countries-main/server/src/models/Country.js
@@ -41,7 +41,7 @@ module.exports = (sequelize) => {
       type: DataTypes.STRING,
     },
     area: {
-      type: DataTypes.FLOAT,
+      type: DataTypes.DECIMAL(10, 1),
     },
     population: {
       type: DataTypes.INTEGER,
